+++ backend/src/main/java/com/shallwecode/backend/user/application/dto/friend/SaveFriendReqDTO.java
@@ -1,4 +1,4 @@
-package com.shallwecode.backend.user.application.dto;
+package com.shallwecode.backend.user.application.dto.friend;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
@@ -8,10 +8,10 @@ import lombok.Setter;
 
 @Getter
 @Setter
-@AllArgsConstructor
 @NoArgsConstructor
-public class UpdateFriendReqDTO {
+@AllArgsConstructor
+public class SaveFriendReqDTO {
 
     @NotNull(message = "Null이면 안됩니다.")
-    private Long fromUserId;
+    private Long toUserId;
 }
